#964 - added TasksList folder, moved some things around
