+++ src/frontend/src/pages/ProjectDetailPage/ProjectViewContainer/TasksList/TaskList.tsx
@@ -8,12 +8,12 @@ import { Box, Button, Tab, Tabs } from '@mui/material';
 import { useEffect, useMemo, useState } from 'react';
 import { Link, useLocation, useRouteMatch } from 'react-router-dom';
 import { Project, Task, TaskStatus, wbsPipe } from 'shared';
-import { useAuth } from '../../../hooks/auth.hooks';
-import PageBlock from '../../../layouts/PageBlock';
-import { routes } from '../../../utils/routes';
-import { Auth } from '../../../utils/types';
+import { useAuth } from '../../../../hooks/auth.hooks';
+import PageBlock from '../../../../layouts/PageBlock';
+import { routes } from '../../../../utils/routes';
+import { Auth } from '../../../../utils/types';
 import TaskListTabPanel from './TaskListTabPanel';
-import LoadingIndicator from '../../../components/LoadingIndicator';
+import LoadingIndicator from '../../../../components/LoadingIndicator';
 
 const TASK_LIST_TITLE: string = 'Task List';
 interface TaskListProps {
