+++ src/frontend/src/pages/ProjectDetailPage/ProjectViewContainer/TasksList/TaskListNotesModal.tsx
@@ -4,7 +4,7 @@
  */
 
 import { TeamPreview, User } from 'shared';
-import { fullNamePipe, datePipe } from '../../../utils/pipes';
+import { fullNamePipe, datePipe } from '../../../../utils/pipes';
 import { Controller, useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import * as yup from 'yup';
@@ -26,11 +26,11 @@ import {
   useTheme
 } from '@mui/material';
 import { DatePicker } from '@mui/x-date-pickers';
-import NERSuccessButton from '../../../components/NERSuccessButton';
-import NERFailButton from '../../../components/NERFailButton';
-import LoadingIndicator from '../../../components/LoadingIndicator';
+import NERSuccessButton from '../../../../components/NERSuccessButton';
+import NERFailButton from '../../../../components/NERFailButton';
+import LoadingIndicator from '../../../../components/LoadingIndicator';
 import { isUnderWordCount, countWords } from 'shared';
-import { useAuth } from '../../../hooks/auth.hooks';
+import { useAuth } from '../../../../hooks/auth.hooks';
 import { useState } from 'react';
 import { Close, Edit } from '@mui/icons-material';
