+++ src/frontend/src/pages/ProjectDetailPage/ProjectViewContainer/TasksList/TaskListTabPanel.tsx
@@ -5,20 +5,20 @@
 
 import { Typography } from '@mui/material';
 import { Box } from '@mui/system';
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { Task, TaskPriority, TaskStatus, UserPreview } from 'shared';
-import LoadingIndicator from '../../../components/LoadingIndicator';
-import { useAuth } from '../../../hooks/auth.hooks';
+import LoadingIndicator from '../../../../components/LoadingIndicator';
+import { useAuth } from '../../../../hooks/auth.hooks';
 import {
   useCreateTask,
   useDeleteTask,
   useEditTask,
   useEditTaskAssignees,
   useSetTaskStatus
-} from '../../../hooks/tasks.hooks';
-import { useToast } from '../../../hooks/toasts.hooks';
-import { TaskListTabPanelProps, transformDate } from '../../../utils/task.utils';
-import ErrorPage from '../../ErrorPage';
+} from '../../../../hooks/tasks.hooks';
+import { useToast } from '../../../../hooks/toasts.hooks';
+import { TaskListTabPanelProps, transformDate } from '../../../../utils/task.utils';
+import ErrorPage from '../../../ErrorPage';
 import TaskListDataGrid from './TaskListDataGrid';
 import TaskListNotesModal, { FormInput } from './TaskListNotesModal';
